Put crates required by test-assembly and test-ui behind feature flag
Means that running a normal `cargo test` doesn't have to build these.
