+++ test-ui/Cargo.toml
@@ -20,9 +20,10 @@ gnustep-1-9 = ["gnustep-1-8", "block2/gnustep-1-9", "objc2-foundation/gnustep-1-
 gnustep-2-0 = ["gnustep-1-9", "block2/gnustep-2-0", "objc2-foundation/gnustep-2-0"]
 gnustep-2-1 = ["gnustep-2-0", "block2/gnustep-2-1", "objc2-foundation/gnustep-2-1"]
 
+run = ["trybuild"]
 
 [dependencies]
-trybuild = "1.0"
+trybuild = { version = "1.0", optional = true }
 block2 = { path = "../block2", default-features = false }
 objc2 = { path = "../objc2", default-features = false }
 objc2-foundation = { path = "../objc2-foundation", default-features = false }
