+++ test-ui/src/main.rs
@@ -2,7 +2,7 @@
 //!
 //! Use as:
 //! ```
-//! TRYBUILD=overwrite cargo +nightly run --bin test-ui
+//! TRYBUILD=overwrite cargo +nightly run --features=run --bin test-ui
 //! ```
 //!
 //! These are not part of the normal test suite, because trybuild doesn't pass
@@ -10,12 +10,17 @@
 //! hence they don't always work.
 //!
 //! Also, they're slower than most tests.
-use std::path::Path;
 
+#[cfg(feature = "run")]
 fn main() {
     let t = trybuild::TestCases::new();
-    let path = Path::new(env!("CARGO_MANIFEST_DIR"))
+    let path = std::path::Path::new(env!("CARGO_MANIFEST_DIR"))
         .join("ui")
         .join("*.rs");
     t.compile_fail(path);
+}
+
+#[cfg(not(feature = "run"))]
+fn main() {
+    panic!("`run` feature must be enabled");
 }
